📦 refactor(tooltip): update module structure and imports
