+++ apps/web/public/components/tooltip/demo/position.ts
@@ -2,7 +2,7 @@ import { CommonModule } from '@angular/common';
 import { Component } from '@angular/core';
 
 import { ZardButtonComponent } from '../../button/button.component';
-import { ZardTooltipModule } from '../tooltip.module';
+import { ZardTooltipModule } from '../tooltip';
 
 @Component({
   standalone: true,
